✨ feat: useInfoModal 커스텀 훅 구현 #93
